Player has health now. Deleted Mario duplicate.
Added a health meter , if it reaches 0 then the game is restarted, maybe I should just restart the level but who knows for now it works. Had to move a bit of the Bub code to make the bullets collision work.

PS I also deleted a repeated mario on a level.