Refactor achievement loading and checking
Instead of loading and checking the achievements on every event, the
achievements are loaded once at the start of the game.

To prevent deeply nested methods a utility functions is added which can
list the files in a directory.

A global achievement list is added for achievements that aren't tied to
a level.

Some of the fields inside of the achievement class have been removed as
they are unecessary and weren't used anywhere.