Refactor achievement loading and add an achievement for shooting ten times (#294)
* Refactor achievement loading and checking

Instead of loading and checking the achievements on every event, the
achievements are loaded once at the start of the game.

To prevent deeply nested methods a utility functions is added which can
list the files in a directory.

A global achievement list is added for achievements that aren't tied to
a level.

Some of the fields inside of the achievement class have been removed as
they are unecessary and weren't used anywhere.

* Restructure achievement folder

There is now a global folder and a level folder which contains numbered
level achievements. The folders have also been converted to lower case.

* Simplify coin achievement file name

Since they are inside the level folder, there is no need to specify the
level inside the file name.

* Add an achievement for shooting ten times

This also adds a new signal to the event bus: "shot". It is emitted when
the player shoots a coin.