Rework localization workflow
Instead of using caps lock keys and translating them into english and
other languages, the english text is now used as a key, which makes it
possible to add translatable text without touching the localization csv file.

This will make it easier to add text-heavy portions of the game, like
dialogs for example.