Add a Github Action that adds generated documentation to the wiki
This pull requests adds a Python file that generates Markdown files from
the source code and pushes it into the Github wiki using a Github
Action.

Only documentation with at least one documented item are added. The
markdown files are added to the "docs" folder, which is regenerated with
every pull request.