Story Mode groundwork: Dialogs Part 4 (#443)
* Add CharacterData resource

Add a new resource that stores the character information. It can be
applied to character instances in the world. Start working on item
pickups. Resize the lever sprite, separate content and logic into
different folders.

* Add YSort to story mode

All characters, the player and items now need to have their origin at
the lowest part of the sprite.

* Add ability to pick up ground items

* Add goto functionality to dialog and fix checks

* Make dialogs progressable using the keyboard

* Allow different characters to speak inside dialogs

* Remove intro

* Keep the speaker if none is specified

* Improve dialog schema descriptions