Add cutscenes to story mode
Cutscene animations are stored inside the CutScenePlayer and are
triggered when an event with the same name is triggered. There is
now a camera that can be used for cutscenes.
The set_sprite dialog option has been removed in favor of setting the
sprite using an animation.

Fixed some warnings.